+++ ansible_collections/studyly/git_mactp/plugins/modules/git_mactp.py
@@ -12,29 +12,29 @@ __metaclass__ = type
 
 DOCUMENTATION = """
 ---
-module: git_acp
+module: git_mactp
 author:
     - "Federico Olivieri (@Federico87)"
 short_description: Perform git add, commit, pull and push operations.
 description:
     - Manage C(git add), C(git commit) C(git push) and C(git pull) on a git repository.
-module: git_acp
+module: git_mactp
 options:
     path:
         description:
             - Folder path where C(.git/) is located.
         type: path
         required: true
-    comment:
+    message:
         description:
-            - Git commit comment. Same as C(git commit -m).
+            - Git commit message. Same as C(git commit -m).
               Required when using C(add).
         type: str
     add:
         description:
             - List of files under C(path) to be staged. Same as C(git add .).
               File globs not accepted, such as C(./*) or C(*).
-              Required when using C(comment).
+              Required when using C(message).
         type: list
         elements: str
         default: ["."]
@@ -43,6 +43,11 @@ options:
             - Git branch where perform git push.
         type: str
         default: main
+    merge:
+        description:
+            - Name of the branch that should be merged into the current branch. Perform merge from given branch into current branch. 
+        type: str
+        version_added: "3.0.0"
     pull:
         description:
             - Perform a git pull before pushing.
@@ -100,6 +105,11 @@ options:
                       C(accept_hostkey)).
                 type: str
         version_added: "1.4.0"
+    tag:
+        description:
+            - Name of the git tag. Tag the current state of the branch.
+        type: str
+        version_added: "3.0.0"
     executable:
         description:
             - Path to git executable to use. If not supplied,
@@ -117,10 +127,10 @@ EXAMPLES = """
     GIT_AUTHOR_EMAIL: "me@me.me"
     GIT_COMMITTER_NAME: "me"
     GIT_COMMITTER_EMAIL: "me@me.me"
-  git_acp:
+  git_mactp:
     path: "{{ working_dir }}"
     branch: "master"
-    comment: "Add {{ file1 }}."
+    message: "Add {{ file1 }}."
     add: [ "." ]
     url: "{{ https_repo }}"
 
@@ -130,10 +140,10 @@ EXAMPLES = """
     GIT_AUTHOR_EMAIL: "me@me.me"
     GIT_COMMITTER_NAME: "me"
     GIT_COMMITTER_EMAIL: "me@me.me"
-  git_acp:
+  git_mactp:
     path: "{{ working_dir }}"
     branch: "master"
-    comment: "Remove {{ file1 }}."
+    message: "Remove {{ file1 }}."
     add: [ "." ]
     url: "{{ https_repo }}"
     push_force: true
@@ -144,8 +154,8 @@ EXAMPLES = """
     GIT_AUTHOR_EMAIL: "me@me.me"
     GIT_COMMITTER_NAME: "me"
     GIT_COMMITTER_EMAIL: "me@me.me"
-  git_acp:
-    comment: "Pull before to push."
+  git_mactp:
+    message: "Pull before to push."
     path: "{{ _pull_dest.path }}"
     url: "{{ _pull_src.path }}"
     pull: true
@@ -156,11 +166,11 @@ EXAMPLES = """
     GIT_AUTHOR_EMAIL: "me@me.me"
     GIT_COMMITTER_NAME: "me"
     GIT_COMMITTER_EMAIL: "me@me.me"
-  git_acp:
+  git_mactp:
     add:
       - "{{ item }}"
     branch: "master"
-    comment: "Add {{ item }}"
+    message: "Add {{ item }}"
     path: "{{ working_dir }}"
     push: false
     url: "{{ https_repo }}"
@@ -174,11 +184,11 @@ EXAMPLES = """
     GIT_AUTHOR_EMAIL: "me@me.me"
     GIT_COMMITTER_NAME: "me"
     GIT_COMMITTER_EMAIL: "me@me.me"
-  git_acp:
+  git_mactp:
     branch: "master"
     path: "{{ working_dir }}"
     url: "{{ https_repo }}"
-    comment: "Add {{ file4 }}"
+    message: "Add {{ file4 }}"
 
 - name: PUSH VIA SSH AND CCEPT_HOSTKEY WHEN SSH DOES NOT SUPPORT THE OPTION
   environment:
@@ -186,11 +196,11 @@ EXAMPLES = """
     GIT_AUTHOR_EMAIL: me@me.me
     GIT_COMMITTER_NAME: me
     GIT_COMMITTER_EMAIL: me@me.me
-  git_acp:
+  git_mactp:
     url: "{{ ssh_repo }}"
     path: "{{ working_dir }}"
     branch: "master"
-    comment: "Remove {{ file2 }}"
+    message: "Remove {{ file2 }}"
     add: [ "{{ file2 }}" ]
     ssh_params:
         accept_hostkey: true
@@ -222,8 +232,7 @@ output:
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.lvrfrc87.git_acp.plugins.module_utils.git_actions import Git
-
+from ansible_collections.studyly.git_mactp.plugins.module_utils.git_actions import Git
 
 def main():
     """
@@ -236,30 +245,41 @@ def main():
             description: returned output from git commands and updated changed status.
     """
     argument_spec = dict(
+        # Path to local repo
         path=dict(required=True, type="path"),
         executable=dict(default=None, type="path"),
-        comment=dict(default=None, type="str"),
-        add=dict(default=".", type="list", elements="str"),
+        message=dict(default=None, type="str"),
+        add=dict(default=None, type="list", elements="str"),
         ssh_params=dict(default=None, type="dict", required=False),
-        branch=dict(default="main"),
+        # Name of working branch
+        branch=dict(default=None),
+        merge=dict(default=None, type="str"),
+        merge_options=dict(default=None, type="list", elements="str"),
         pull=dict(default=False, type="bool"),
         pull_options=dict(default=["--no-edit"], type="list", elements="str"),
         push=dict(default=True, type="bool"),
         push_option=dict(default=None, type="str"),
         push_force=dict(default=False, type="bool"),
+        # Name of tag
+        tag=dict(default=None, type="str"),
         url=dict(required=True, no_log=True),
     )
 
     module = AnsibleModule(
         argument_spec=argument_spec,
-        required_together=[("comment", "add")],
-        required_one_of=[("add", "pull", "push")]
+        # required_together=[("message", "add")],
+        required_one_of=[("add", "pull", "push", "tag", "branch", "merge")],
+        mutually_exclusive=["add", "tag", "merge"]
     )
 
+    add = module.params.get("add")
     url = module.params.get("url")
+    merge = module.params.get("merge")
     pull = module.params.get("pull")
     push = module.params.get("push")
+    branch = module.params.get("branch")
     ssh_params = module.params.get("ssh_params")
+    tag = module.params.get("tag")
 
     module.run_command_environ_update = dict(
         LANG="C.UTF-8", LC_ALL="C.UTF-8", LC_MESSAGES="C.UTF-8", LC_CTYPE="C.UTF-8"
@@ -277,18 +297,37 @@ def main():
             )
 
     git = Git(module)
+    # checkout branch if parameter was given
+    if branch:
+        git.checkout()
+
+    if pull:
+        result.update(git.pull())
+    
     changed_files = git.status()
 
-    if changed_files:
-        if pull:
-            result.update(git.pull())
-
+    if merge:
+        result.update(git.merge())
+    elif changed_files and add:
         git.add()
         result.update(git.commit())
+    elif tag:
+        result.update(git.tag())
+    else:
+        push = False
 
-        if push:
-            result.update(git.push())
-        result["changed"] = True
+    if push:
+        result.update(git.push())
+
+    result["changed"] = False
+
+    for d in result:
+        if type(d) == dict:
+            if "changed" in d:
+                if d["changed"]:
+                    result["changed"] = True
+                    break
+         
 
     module.exit_json(**result)