+++ deployment/dev.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-cd ../ansible_collections/lvrfrc87/git_acp/ && \
+cd ../ansible_collections/studyly/git_mactp/ && \
 ansible-test sanity --docker -v plugins/ && \
 ansible-test integration -v --docker && \
-ansible-lint tests/integration/targets/git_acp/tasks/main.yml
+ansible-lint tests/integration/targets/git_mactp/tasks/main.yml